Cleaning up.
Specifically, removed some files im never going to bother with, added the base for the api handler, and changed the installation file to save a username and password. (Now i just need to DO something with the username and pass from notes)
Also worked on cleaning up the notes file entirely.